requirements.txt
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
click==8.1.8
exceptiongroup==1.3.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
